evaluate_data/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
alembic==1.14.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi>=2023.7.22
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.0.1
colorlog==6.9.0
Cython==0.29.28
deepspeed==0.16.4
docker-pycreds==0.4.0
easydict==1.9
einops==0.8.1
eval_type_backport==0.2.2
frozenlist==1.5.0
gdown==5.2.0
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
greenlet==3.1.1
grpcio==1.70.0
h5py==3.6.0
hjson==3.1.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
joblib==1.4.2
lightning-utilities==0.11.9
lit==18.1.8
Mako==1.3.9
matplotlib==3.5.1
mkl-service==2.4.0
msgpack==1.1.0
multidict==6.1.0
multimethod==1.7
munkres==1.1.4
ninja==1.10.2.3
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
packaging==23.1
pandas==1.4.2
patsy ==0.5.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==4.3.6
pooch==1.8.2
propcache==0.2.0
protobuf==3.19.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.9.3
pytz==2023.3
pyvista==0.44.2
PyYAML==6.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.0.2
scooby==0.10.0
seaborn==0.11.2
sentry-sdk==2.23.1
setproctitle==1.3.5
six==1.16.0
smmap==5.0.2
soupsieve==2.6
SQLAlchemy==2.0.39
statsmodels==0.13.2
svgwrite==1.4.3
sympy==1.11.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.5.0
tqdm==4.62.3
triton==2.0.0
typing_extensions==4.12.2
tzdata==2025.1
unicodedata2==14.0.0
urllib3==2.2.3
vtk==9.3.1
wandb==0.19.8
Werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
